Fix AdminEngine
* move a routing from config/routes into AdminEngine.routes
* fix path of partial file
* fix locale file
