+++ config/locales/ja.yml
@@ -53,22 +53,6 @@ ja:
       officializations:
         index:
           show_user_extension: ユーザー属性を表示
-        show_user_extension_modal:
-          user_extension: ユーザー属性
-          close_modal: モーダルを閉じる
-          description: 直接参加者に連絡する必要がある場合は，表示ボタンをクリックしてメールアドレスを見ることができます．このアクションは記録されます．
-          email_address: メールアドレス
-          full_name: フルネーム
-          hidden: hidden
-          show: 表示
-          title: 参加者の属性情報を表示
-        user_extensions:
-          show:
-            real_name: 本名
-            address: 住所
-            birth_year: 生年
-            gender: 性別
-            occupation: 職業
     components:
       components:
         comment:
@@ -185,6 +169,25 @@ ja:
           deleted: 退会者
         sortition_cancel_author:
           deleted: 退会者
+    user_extension:
+      admin:
+        officializations:
+          user_extensions:
+            show:
+              real_name: 本名
+              address: 住所
+              birth_year: 生年
+              gender: 性別
+              occupation: 職業
+          show_user_extension_modal:
+            user_extension: ユーザー属性
+            close_modal: モーダルを閉じる
+            description: 直接参加者に連絡する必要がある場合は，表示ボタンをクリックしてメールアドレスを見ることができます．このアクションは記録されます．
+            email_address: メールアドレス
+            full_name: フルネーム
+            hidden: hidden
+            show: 表示
+            title: 参加者の属性情報を表示
     verifications:
       postal_letter:
         admin:
