+++ config/routes.rb
@@ -8,10 +8,5 @@ Rails.application.routes.draw do
     mount Sidekiq::Web => "/sidekiq"
   end
 
-  get "/admin/officializations/user_extensions/:user_id" =>
-      "decidim/admin/officializations/user_extensions#show",
-      constraints: (->(request) { Decidim::Admin::OrganizationDashboardConstraint.new(request).matches? }),
-      as: "user_extension_admin_officialization"
-
   mount LetterOpenerWeb::Engine, at: "/letter_opener" if Rails.env.development?
 end
