+++ decidim-user_extension/app/controllers/decidim/user_extension/admin/application_controller.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Decidim
+  module UserExtension
+    module Admin
+      class ApplicationController < Decidim::Admin::ApplicationController
+      end
+    end
+  end
+end
