+++ decidim-user_extension/app/overrides/decidim/admin/officializations/index/user_extension_modal_override.html.erb.deface
@@ -1,2 +1,2 @@
 <!-- insert_after "div#user-groups" -->
-<%= render "show_user_extension_modal" %>
+<%= render "decidim/user_extension/admin/officializations/show_user_extension_modal" %>
