+++ decidim-user_extension/app/overrides/decidim/admin/officializations/index/user_extension_override.html.erb.deface
@@ -1,5 +1,5 @@
 <!-- insert_bottom "td.table-list__actions" -->
 <% if allowed_to? :show_user_extension, :user, user: user %>
-  <%= icon_link_to "person", main_app.user_extension_admin_officialization_path(user.id), t(".show_user_extension"), class: "action-icon action-icon--show-user", data: { full_name: user.name, toggle: "show-user-modal" } %>
+  <%= icon_link_to "person", Decidim::UserExtension::AdminEngine.routes.url_helpers.officializations_show_user_extension_path(user.id), t(".show_user_extension"), class: "action-icon action-icon--show-user", data: { full_name: user.name, toggle: "show-user-modal" } %>
 <% end %>
