+++ decidim-user_extension/lib/decidim/user_extension.rb
@@ -1,5 +1,6 @@
 require "decidim/user_extension/admin"
 require "decidim/user_extension/engine"
+require "decidim/user_extension/admin_engine"
 
 module Decidim
   module UserExtension
